requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
autopep8==2.0.4
certifi==2023.7.22
charset-normalizer==3.2.0
dataclasses-json==0.6.1
diskcache==5.6.3
docopt==0.6.2
docutils==0.20.1
exceptiongroup==1.1.3
faiss-cpu==1.7.4
FLAML==2.1.0
frozenlist==1.4.0
idna==3.4
importlib-metadata==6.8.0
jaraco.classes==3.3.0
jsonpatch==1.33
jsonpointer==2.4
keyring==24.2.0
langchain==0.0.305
langsmith==0.0.41
markdown-it-py==3.0.0
marshmallow==3.20.1
mdurl==0.1.2
more-itertools==10.1.0
multidict==6.0.4
mypy-extensions==1.0.0
nh3==0.2.14
numexpr==2.8.7
numpy==1.26.0
openai==0.28.1
packaging==23.1
pkginfo==1.9.6
pyautogen==0.1.4
pycodestyle==2.11.0
pydantic==2.4.2
pydantic_core==2.10.1
Pygments==2.16.1
python-dotenv==1.0.0
PyYAML==6.0.1
readme-renderer==42.0
regex==2023.8.8
requests==2.31.0
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==13.6.0
sniffio==1.3.0
SQLAlchemy==2.0.21
tenacity==8.2.3
termcolor==2.3.0
tiktoken==0.5.1
tomli==2.0.1
tqdm==4.66.1
twine==4.0.2
typing-inspect==0.9.0
typing_extensions==4.8.0
urllib3==2.0.5
websockets==11.0.3
yarl==1.9.2
zipp==3.17.0
